Refactor appointment management to implement ReservarCitaUseCase, update Cita model, and introduce Ticket generation

+++ src/main/java/org/hearelmer/citas/infrastructure/persistence/jpa/repository/SpringDataCitaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpringDataCitaRepository extends JpaRepository<CitaJpaEntity, Long> {
+public interface SpringDataCitaRepository extends JpaRepository<CitaJpaEntity, Integer> {
 }
